# contracts/public-api.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ProofRun Case API",
    "version": "0.1.0",
    "description": "新接入使用 /v2/cases 批量提交结构化步骤，/v1/cases 保留兼容。环境、身份、预算及执行策略由平台管理。v2 每步独立 300 秒、10 次动作、20 次模型请求（JEV 40 次），耗尽后继续后续步骤，并双跑纯 LLM 与 JEV+LLM；comparison.arms 返回各组结果，顶层仅表示主组。接口需要部署管理员 Bearer 凭据，请仅由可信服务端保管；提交和取消会产生实际操作。由共享 Schema 生成，请勿手改。"
  },
  "servers": [
    {
      "url": "https://api-proofrun.ethankit.com",
      "description": "生产 HTTP API"
    },
    { "url": "/", "description": "自部署时替换为实际控制面 origin。" }
  ],
...
  "security": [{ "AdminBearer": [] }],
...
  "paths": {
    "/v2/cases": {
      "post": {
        "operationId": "submitCasesV2",
        "tags": ["Case API v2"],
        "summary": "原子接收结构化 case 数组；每个对象是独立任务，内部串行",
        "responses": {
          "202": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/CaseResultV2" }
                }
              }
            }
          },
          "default": { "$ref": "#/components/responses/Error" }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "minItems": 1,
                "maxItems": 32,
                "items": { "$ref": "#/components/schemas/VerificationCaseV2" }
              }
            }
          }
        }
      }
    },
    "/v2/cases/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": { "type": "string" }
        }
      ],
      "get": {
        "operationId": "getCaseV2",
        "tags": ["Case API v2"],
        "summary": "读取步骤、验收与关联清理结果",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CaseResultV2" }
              }
            }
          },
          "default": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/v2/cases/{id}/rerun": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": { "type": "string" }
        }
      ],
      "post": {
        "operationId": "rerunCaseV2",
        "tags": ["Case API v2"],
        "summary": "使用新 caseId 重跑整个 case；默认采用当前预算策略，可选择原预算",
        "responses": {
          "202": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CaseResultV2" }
              }
            }
          },
          "default": { "$ref": "#/components/responses/Error" }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "required": ["caseId"],
                "properties": {
                  "caseId": {
                    "type": "string",
                    "pattern": "^[A-Za-z0-9_-]{1,80}$"
                  },
                  "budgetMode": {
                    "enum": ["current", "original"],
                    "default": "current"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v2/cases/{id}/cancel": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": { "type": "string" }
        }
      ],
      "post": {
        "operationId": "cancelCaseV2",
        "tags": ["Case API v2"],
        "summary": "取消主任务；已声明的副作用清理仍独立执行",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CaseResultV2" }
              }
            }
          },
          "default": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/v2/cases/{id}/evidence/{artifactId}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": { "type": "string" }
        },
        {
          "name": "artifactId",
          "in": "path",
          "required": true,
          "schema": { "type": "string" }
        }
      ],
      "get": {
        "operationId": "downloadCaseEvidenceV2",
        "tags": ["Case API v2"],
        "summary": "读取主任务或关联清理结果的证据",
        "responses": {
          "200": {
            "description": "截图为 PNG，其他证据为 JSON。",
            "content": {
              "image/png": {
                "schema": { "type": "string", "format": "binary" }
              },
              "application/json": {
                "schema": { "type": "object", "additionalProperties": true }
              }
            }
          },
          "default": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/v1/cases": {
      "post": {
        "operationId": "submitCase",
        "tags": ["Case API v1（兼容）"],
        "summary": "幂等提交一次 case 验证",
        "responses": {
          "202": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CaseResult" }
              }
            }
          },
          "default": { "$ref": "#/components/responses/Error" }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/VerificationCase" },
              "example": {
                "caseId": "notification-default-001",
                "description": "验证通知设置中邮件通知的默认状态。只检查，不修改或保存。",
                "url": "https://app.example.com/settings",
                "steps": ["进入通知设置区域", "观察邮件通知开关，不点击开关"],
                "acceptanceCriteria": [
                  {
                    "id": "email-enabled",
                    "description": "在通知设置区域读取邮件通知开关状态。",
                    "expectedResult": "邮件通知默认开启。"
                  }
                ]
              }
            }
          }
        }
      }
    },
    "/v1/cases/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": { "type": "string" }
        }
      ],
      "get": {
        "operationId": "getCase",
        "tags": ["Case API v1（兼容）"],
        "summary": "读取 case 状态和结果",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CaseResult" }
              }
            }
          },
          "default": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/v1/cases/{id}/cancel": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": { "type": "string" }
        }
      ],
      "post": {
        "operationId": "cancelCase",
        "tags": ["Case API v1（兼容）"],
        "summary": "取消本次验证；已完成的业务操作不回滚",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CaseResult" }
              }
            }
          },
          "default": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/v1/cases/{id}/evidence/{artifactId}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": { "type": "string" }
        },
        {
          "name": "artifactId",
          "in": "path",
          "required": true,
          "schema": { "type": "string" }
        }
      ],
      "get": {
        "operationId": "downloadCaseEvidence",
        "tags": ["Case API v1（兼容）"],
        "summary": "读取本 case 报告引用的证据",
        "responses": {
          "200": {
            "description": "截图为 PNG，其他证据为 JSON。",
            "content": {
              "image/png": {
                "schema": { "type": "string", "format": "binary" }
              },
              "application/json": {
                "schema": { "type": "object", "additionalProperties": true }
              }
            }
          },
          "default": { "$ref": "#/components/responses/Error" }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "AdminBearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "PROOFRUN_ADMIN_TOKEN；当前为全局管理员权限。"
      }
    },
    "responses": {
      "Error": {
        "description": "应用错误；按 HTTP 状态和 code 处理，网关错误可能不是此结构。",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" }
          }
        }
      }
    },
    "schemas": {
...
      "VerificationCase": {
        "title": "VerificationCase",
        "description": "上层定义的单次验证 case；环境、登录态、取证方式和预算由平台配置。",
        "type": "object",
        "additionalProperties": false,
        "required": ["caseId", "description", "url", "acceptanceCriteria"],
        "properties": {
          "caseId": {
            "type": "string",
            "pattern": "^[A-Za-z0-9_-]{1,100}$",
            "description": "本次验证的幂等身份；再次执行使用新的身份。"
          },
          "description": {
            "type": "string",
            "minLength": 1,
            "pattern": "\\S",
            "description": "完整任务描述：验证目标、业务背景、涉及对象及必要约束。"
          },
          "url": {
            "type": "string",
            "format": "uri",
            "pattern": "^https?://",
            "description": "本次验证的 HTTP(S) 入口地址，由调用方明确提供，必须能从执行节点访问。"
          },
          "steps": {
            "type": "array",
            "minItems": 1,
            "maxItems": 32,
            "items": {
              "type": "string",
              "minLength": 1,
              "pattern": "\\S",
              "maxLength": 2000
            },
            "description": "可选的必要业务步骤；不要求提供选择器或固定页面点击路径。"
          },
          "acceptanceCriteria": {
            "type": "array",
            "minItems": 1,
            "items": {
              "type": "object",
              "additionalProperties": false,
              "required": ["id", "description", "expectedResult"],
              "properties": {
                "id": {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 128,
                  "pattern": "\\S",
                  "description": "任务内唯一的验收项身份，结果通过 criterionId 对应。"
                },
                "description": {
                  "type": "string",
                  "minLength": 1,
                  "pattern": "\\S",
                  "description": "检查对象、范围及必要观察时机。"
                },
                "expectedResult": {
                  "type": "string",
                  "minLength": 1,
                  "pattern": "\\S",
                  "description": "可以据证据判断的预期业务结果。"
                }
              }
            },
            "description": "上层定义的完整业务验收项；缺少标准时不由平台自行补写。"
          }
        }
      },
...
      "VerificationCaseV2": {
        "title": "VerificationCaseV2",
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "caseId": {
            "type": "string",
            "pattern": "^[A-Za-z0-9_-]{1,80}$",
            "description": "本次任务的幂等身份；重跑必须使用新的身份。"
          },
          "platform": {
            "type": "string",
            "minLength": 1,
            "pattern": "\\S",
            "maxLength": 2000,
            "description": "业务平台标签，不用于自动选择登录账号。"
          },
          "entry": {
            "type": "string",
            "format": "uri",
            "pattern": "^https?://",
            "description": "任务所属站点的 HTTP(S) 入口。"
          },
          "description": {
            "type": "string",
            "minLength": 1,
            "pattern": "\\S",
            "maxLength": 2000,
            "description": "任务背景，省略时由平台生成不含额外验收标准的概述。"
          },
          "steps": {
            "type": "array",
            "minItems": 1,
            "maxItems": 32,
            "items": { "$ref": "#/components/schemas/CaseStep" },
            "contains": {
              "type": "object",
              "properties": { "type": { "const": "verification" } },
              "required": ["type"]
            },
            "description": "上游定义的操作与验收步骤；服务端补齐身份后稳定排序。"
          },
          "cleanup": {
            "type": "array",
            "maxItems": 32,
            "items": {
              "type": "object",
              "additionalProperties": false,
              "properties": {
                "stepId": {
                  "type": "string",
                  "pattern": "^[A-Za-z0-9_-]{1,80}$"
                },
                "url": {
                  "type": "string",
                  "format": "uri",
                  "pattern": "^https?://"
                },
                "exec_order": {
                  "type": "integer",
                  "minimum": 1,
                  "maximum": 1000000,
                  "description": "升序执行；同值按原数组顺序串行。"
                },
                "description": {
                  "type": "string",
                  "minLength": 1,
                  "pattern": "\\S",
                  "maxLength": 2000
                },
                "policy": {
                  "type": "array",
                  "maxItems": 32,
                  "items": {
                    "type": "string",
                    "minLength": 1,
                    "pattern": "\\S",
                    "maxLength": 2000
                  }
                },
                "expected": {
                  "type": "array",
                  "maxItems": 32,
                  "items": {
                    "type": "string",
                    "minLength": 1,
                    "pattern": "\\S",
                    "maxLength": 2000
                  }
                }
              },
              "required": ["url", "exec_order", "description"]
            },
            "description": "在同一任务和浏览器会话末尾执行的业务清理操作，不是浏览器资源回收。"
          }
        },
        "required": ["caseId", "platform", "entry", "steps", "cleanup"]
      },
...
      "Error": {
        "type": "object",
        "additionalProperties": false,
        "required": ["code", "message"],
        "properties": {
          "code": { "type": "string" },
          "message": { "type": "string" }
        }
      }
    }
  }
}